laptop changes

+++ install.sh
@@ -9,7 +9,13 @@ if ! command -v paru &> /dev/null; then
     rm -r paru
 fi
 
-paru --needed -S - < pkglist.txt
+read -p "Do you want to update packages? (yes/no): " choice
+
+if [[ "$choice" == "yes" ]]; then
+    paru --needed -S - < pkglist.txt
+else
+    echo "Skipping package update."
+fi
 
 
 
@@ -21,11 +27,28 @@ systemctl --user enable swaync.service
 
 go install github.com/nehrbash/hyprshell@latest
 
-chsh -s $(which zsh)
+if [[ "$SHELL" == *"/zsh" ]]; then
+    echo "The current shell is already zsh. Skipping shell change."
+else
+    read -p "Do you want to change the shell to zsh? (yes/no): " choice
 
-PROFILE=$(awk -F= -v section="$install_section" '$1 == "Default" && found {print $2; exit} $1 == section {found=1}' ~/.mozilla/firefox/profiles.ini)
-# Create the symlink
-ln -sfn ~/.dotfiles/.config/chrome/ ~/.mozilla/firefox/${PROFILE}/
+    if [[ "$choice" == "yes" ]]; then
+        chsh -s $(which zsh)
+        echo "Shell changed to zsh."
+    else
+        echo "Skipping shell change."
+    fi
+fi
+
+read -p "Do you want to update Firefox CSS? (yes/no): " choice
+
+if [[ "$choice" == "yes" ]]; then
+	local PROFILE=$(awk -F= -v section="$install_section" '$1 == "Default" && found {print $2; exit} $1 == section {found=1}' ~/.mozilla/firefox/profiles.ini)
+	# Create the symlink
+	ln -sfn ~/.dotfiles/.config/chrome/ ~/.mozilla/firefox/${PROFILE}/
+else
+    echo "Skipping Firefox CSS update."
+fi
 
 spicetify config current_theme Onepunch color_scheme light
 spicetify backup
